devide for 2 tab
