+++ client-side/src/App.jsx
@@ -6,13 +6,17 @@ import { router } from './router/router.jsx';
 import { store } from './redux/store.jsx';
 import Header from './stories/header/header'
 import Footer from './stories/footer/FooterComponent'
+import Settings from './components/settings/Notifications.jsx';
 
 function App() {
   
+   
+
   
   return (
 
     <>
+      {/* <RouterProvider router={router} /> */}
       <Provider store={store}>
       <Header/>
         <Footer />
