+++ client-side/src/components/settings/Notifications.jsx
@@ -5,17 +5,16 @@ import GenericButton  from '../../stories/Button/GenericButton.jsx';
 import GenericInput from '../../stories/GenericInput/genericInput.jsx'
 import Select from '../../stories/Select/Select.jsx'
 import {uploadFile} from './uploadFileUtil.js'
+import Preferences from './preferences.jsx';
 import {LANGUAGE,LABELS} from './constantSetting.js'
-const Settings = ({currentUser={}}) => {
+const Notifications = ({currentUser={}}) => {
     const {emailFrequency,sendNotificationTime,_id,soundVoice}= currentUser.preference
     const url=process.env.REACT_APP_BASE_URL;
     const preferencesUrl=`${url}/preferences/${_id}`
     const [ringtoneFile, setRingtoneFile] = useState(null);
     const [audioSrc,setAudioSrc]  = useState(`${url}/uploads/${soundVoice}`);
     const [notificationTime,setNotificationTime]= useState(sendNotificationTime);
-    const [lng,setLng] =useState('en');
-    const {t,i18n}= useTranslation();
-
+    const {t}= useTranslation();
     const handleFileChange=(e) => {
          if(e.target.files[0])
          { 
@@ -25,12 +24,7 @@ const Settings = ({currentUser={}}) => {
           
     };
     
-     const handleLngChange=(value)=>{
-      console.log('value= ', value);
-       i18n.changeLanguage(value);
-          setLng(value)
-     }
-   
+    
      const changeNotificationTime=(event)=>{
          setNotificationTime(event);
      }
@@ -57,27 +51,17 @@ const Settings = ({currentUser={}}) => {
             
           </div>       
 
-         <Select  title={t(LABELS.SELECT_LANGUAGES)} 
-          options={Object.keys(LANGUAGE).map(key=>({
-            value:key,
-            text:LANGUAGE[key]['text'],
-            iconSrc:LANGUAGE[key]['icon']
-          }))} 
-          className='select-class' 
-          size={'large'}
-          widthOfSelect='200px'
-          onChange={handleLngChange}/>
+         
+          <Preferences/>
         </div>          
-        
-          <GenericInput size='small'  label={t(LABELS.CHANGE_NOTIFICATION_TIME)} onChange={changeNotificationTime} type='number' className='' />
-          <GenericButton size='small'  label={t(LABELS.SEND_PREFERENCE)} onClick={sendPreference} className='but-send'/>
-
+         <GenericInput size='small'  label={t(LABELS.CHANGE_NOTIFICATION_TIME)} onChange={changeNotificationTime} type='number' className='' />
+         <GenericButton size='small'  label={t(LABELS.SEND_PREFERENCE)} onClick={sendPreference} className='but-send'/>
 
       </>
          
     );
 };
- Settings.propTypes={
+Notifications.propTypes={
     currentUser: PropTypes.object.isRequired
  }
- export default Settings;
+ export default Notifications;
